requirements.txt
python-dotenv>=1.0.0
yfinance>=0.2.40
pandas>=2.0.0
python-telegram-bot>=20.0
pytz>=2024.1
requests>=2.0.0